refactor: move types to src
