+++ plugins/quarkus-console/src/types.ts
